msm8610-common: Force NV12 and NV21 pixel alignment
New versions of libadreno_utils return an incorrect value for
the pixel alignment (32 instead of 16) and therefore the 480p
preview is corrupted. Force the right alignment to fix it.

This change allows us to use the proper Marshmallow CAF Adreno
drivers rather than being forecd to use buggy older drivers
that lack the function to get pixel alignment.

Change-Id: Ia37a39b9582d2137006c609be2c24b72a7f5fa0f